Clean up NewMarketStep components
